% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFuns.R
\name{calcWeights}
\alias{calcWeights}
\title{Calculate observation weights}
\usage{
calcWeights(residuals)
}
\arguments{
\item{residuals}{(numeric) Residuals from a (generalized additive) model fit}
}
\value{
Estimated observation weights, giving no or reduced weight to
outlying observations.
}
\description{
Calculate observation weights
}
\details{
This function calculates weights based on the residuals of a model fit, here
used for generalized additive model fits. Residuals that are large are given
lower or zero weight, suggesting that these are outliers and should
contribute less to the model fit. The intention is to obtain fits robust to
outliers and overfitting.

The weighting approach is Tukey's biweight approach.

Residuals may include NA values if observations were not used for fitting
the initial model. A zero weight will be returned for these cases.
}
