% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{curate_data}
\alias{curate_data}
\title{Meta analysis data curation}
\usage{
curate_data(dataset, ...)
}
\arguments{
\item{dataset}{This can either be a data.frame or the filename to the dataset}

\item{...}{Other paramters to be used by the funtion read.csv}
}
\value{
A dataframe with curated data
}
\description{
This function curates dataset for meta analysis
}
