% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_regression_coefs}
\alias{find_regression_coefs}
\title{Find regression coefficients of each patient}
\usage{
find_regression_coefs(df)
}
\arguments{
\item{df}{A curated data.frame}
}
\value{
A list that contains the data.frame with residuals and a data.frame containing all regression coefficients
}
\description{
This function finds regression coefficients of each patient
}
