% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fitWithPosCtrl}
\alias{fitWithPosCtrl}
\title{Find the simple linear regression coefficients}
\usage{
fitWithPosCtrl(y, x)
}
\arguments{
\item{y}{A vector of responses}

\item{x}{A vector of features}
}
\value{
The coefficients of the fitted simple linear regression
}
\description{
Given the response y and the feature x, this function
fits a simple linear regression and returns its coefficients
}
