% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_range}
\alias{get_range}
\title{Get prior range of a uniform distribution}
\usage{
get_range(x, mm = 5)
}
\arguments{
\item{x}{A vector}

\item{mm}{Number of standard deviations}
}
\value{
The lower and upper bounds of a uniform distribution
}
\description{
This funciton gets a prior range of a uniform distribution
This is mainly used for setting up the priors for the Gibbs sampler
}
