% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_residual}
\alias{get_residual}
\title{Get residual from positive control fitted with simple linear regression}
\usage{
get_residual(log_dat, RNA_conc, coefs)
}
\arguments{
\item{RNA_conc}{A array of RNA concentrations}

\item{coefs}{A matrix of coefficients of simple linear regressions}

\item{log_data}{A matrix of log-transformed data}
}
\value{
A matrix of residuals
}
\description{
This function extracts residuals from a series of fitted
simple linear regressions
}
