% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inverse}
\alias{inverse}
\title{Compute matrix inverse}
\usage{
inverse(sig)
}
\arguments{
\item{sig}{A matrix to be inversed}
}
\value{
The inverse of the supplied matrix
}
\description{
This function computes the inverse of the supplied matrix
}
