% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{meta_analysis}
\alias{meta_analysis}
\title{Perform meta analysis}
\usage{
meta_analysis(ds, coeffs2, M = 5000, n_keep = -1)
}
\arguments{
\item{ds}{A curated data.frame}

\item{coeffs2}{A data.frame of estimated coefficients}

\item{M}{Number of draws}

\item{n_keep}{Number of samples to keep}
}
\value{
A data.frame with posterior draws
}
\description{
This function performs meta analysis via Gibbs sampling
}
