% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_low_memory.R
\name{MetaNorm_LowMemory}
\alias{MetaNorm_LowMemory}
\title{Estimate mRNA amount for every gene/sample combination}
\usage{
MetaNorm_LowMemory(dat, M = 15000, burn_in = 5000, thin = 1, mm = 3, seed = 0)
}
\arguments{
\item{dat}{A list of 4 dataframes}

\item{M}{Number of MCMC draws}

\item{burn_in}{Number of burn in MCMC draws}

\item{thin}{Number of thinning in MCMC draws}

\item{mm}{The radius}

\item{seed}{Random seed}
}
\value{
A list containing two items: “GlobalParameters_Draws” (all draws for mu_a, mu_b, sig2_a, sig2_b) and “Posterior_estimates” (list of kappa_hk, kappa_reg posterior estimates)
}
\description{
This function is similar to RCRnorm with the prior changed
to an informative prior infered from the meta analysis. The underlying
model and inference procedure is exactly the same as the MetaNorm function.
However, it only keeps the most recent draws for each parameter except for
global parameters whose posterior means are computed via a rolling mean method.
}
