% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mu_cpp}
\alias{mu_cpp}
\title{A helper function of the Gibbs sampler to mu the parameters a_{ik} and b_{ik}}
\usage{
mu_cpp(Sig1, Sig, X, sigjk, m, Y, s)
}
\arguments{
\item{Sig1}{The updated posterior covariance matrix}

\item{Sig}{The current posterior sample of the covariance matrix}

\item{X}{X A matrix of X_{ik}}

\item{sigjk}{A diagonal matrix of the posterior samples of sigma^2_{jk}}

\item{m}{The current posterior sample of the mean vector}

\item{Y}{A vector of Y_{ik}}

\item{s}{A vector of s_k}
}
\value{
An updated mu vector
}
\description{
This function computes the posterior mean vector
of the parameter theta_{ik} = (a_{ik} and b_{ik})
}
