% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFuns.R
\name{plotPCA}
\alias{plotPCA}
\title{Generate a PCA plot of a data matrix, coloring observations by a factor
variable 'type'}
\usage{
plotPCA(x, type)
}
\arguments{
\item{x}{(matrix) Numerical matrix, columns containing samples, and rows
containing compounds}

\item{type}{(character) Type of sample, e.g., QCs vs. non QCs, or different
batches.}
}
\value{
A ggplot
}
\description{
Generate a PCA plot of a data matrix, coloring observations by a factor
variable 'type'
}
