% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rMvNorm}
\alias{rMvNorm}
\title{Generate random variables from a multivariate Gaussian distribution}
\usage{
rMvNorm(n, mu, sigma)
}
\arguments{
\item{n}{The number of samples}

\item{mu}{The mean vector}

\item{sigma}{The variance-covariance matrix}
}
\value{
A matrix of random draws from the specified Gaussian distribution
}
\description{
This function uses the Cholesky decomposition to transform
isotropic Gaussian samples to samples from the desired Gaussian distribution
}
