% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFuns.R
\name{selectSpan}
\alias{selectSpan}
\title{LOESS span selection with LOOCV / MSE as criterion}
\usage{
selectSpan(x, y, method = c("QC-RLSC", "rLOESS"))
}
\arguments{
\item{x}{(numeric) Injection order}

\item{y}{(numeric) Observed intensities}

\item{method}{(character) Either "QC-RLSC" or "rLOESS"}
}
\value{
The optimal span for fitting a (robust) LOESS model.
}
\description{
LOESS span selection with LOOCV / MSE as criterion
}
