% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sig1_cpp}
\alias{Sig1_cpp}
\title{A helper function of the Gibbs sampler to the Sigma parameters of a_{ik} and b_{ik}}
\usage{
Sig1_cpp(Sig, X, sigjk)
}
\arguments{
\item{Sig}{The current posterior sample of the covariance matrix}

\item{X}{A matrix of X_{ik}}

\item{sigjk}{A diagonal matrix of the posterior samples of sigma^2_{jk}}
}
\value{
An updated Sigma matrix
}
\description{
This function computes the posterior covariance matrix
of the parameter theta_{ik} = (a_{ik} and b_{ik})
}
