% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{sum_except}
\alias{sum_except}
\title{Vector version of summing up every thing except for the ith element}
\usage{
sum_except(i, vec)
}
\arguments{
\item{i}{The index to omit}

\item{vec}{The vector}
}
\value{
The sum
}
\description{
Vector version of summing up every thing except for the ith element
}
