% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{sum_except2}
\alias{sum_except2}
\title{Sum up every thing except for the ith element}
\usage{
sum_except2(i, vec)
}
\arguments{
\item{i}{The index to omit}

\item{vec}{The vector}
}
\value{
The sum
}
\description{
Sum up every thing except for the ith element
}
