% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_gibbs.R
\name{update_ab}
\alias{update_ab}
\title{Update parameter a_{ik} and b_{ik}}
\usage{
update_ab(i, k, indx, m, Sig, sigjk, sjk)
}
\arguments{
\item{i}{The patient index}

\item{k}{The study index}

\item{indx}{The prob index}

\item{m}{The current posterior sample of the mean}

\item{Sig}{The current posterior sample of the covariance}

\item{sigjk}{The current posterior sample of the variances of the residuals}

\item{sjk}{The current posterior sample of s_{jk}}
}
\value{
A sample of the \theta parameter
}
\description{
This function performs one step of the Gibbs sampling
of the parameter \theta_{ik} = (a_{ik}, b_{ik})
}
