% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_gibbs.R
\name{update_m}
\alias{update_m}
\title{Update parameter m}
\usage{
update_m(k, ab, sig_alpha, sig_beta, Sig, mu)
}
\arguments{
\item{k}{The study index}

\item{ab}{The current posterior sample of the intercept and the slope}

\item{sig_alpha}{The current posterior sample of sigma^2_alpha}

\item{sig_beta}{The current posterior sample of sigma^2_beta}

\item{Sig}{The current posterior sample of the covarance}

\item{mu}{The current posterior sample of the mu parameter}
}
\value{
A sample of the m parameter
}
\description{
This function performs one step of the Gibbs sampling
of the parameter m
}
