% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_gibbs.R
\name{update_mu}
\alias{update_mu}
\title{Update parameter mu}
\usage{
update_mu(L, U, alpha, sig_alpha, K = 13)
}
\arguments{
\item{L}{The lower bound}

\item{U}{The upper bound}

\item{alpha}{The current posterior sample of alpha/beta}

\item{sig_alpha}{The current posterior sample of sigma^2_alpha/sigma^2_beta}

\item{K}{The number of studies}
}
\value{
A sample of the mu_alpha/mu_beta parameter
}
\description{
This function performs one step of the Gibbs sampling
of the parameter mu. As the form of the updating rules
are identical for mu_alpha and mu_beta, we use the same
function for both of them
}
