% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_a}
\alias{update_norm_a}
\title{Update a}
\usage{
update_norm_a(
  Y,
  X_pos,
  Y_neg,
  Y_hk,
  Y_reg,
  cc,
  b,
  d_pos,
  d_neg,
  d_hk,
  d_reg,
  mu_a,
  phi,
  kappa_hk,
  kappa_reg,
  sig2_a,
  sig2_e,
  sig2_n,
  P,
  N,
  H,
  R,
  n_samples
)
}
\value{
A posterior sample of a
}
\description{
Update a
}
