% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_b}
\alias{update_norm_b}
\title{Update b}
\usage{
update_norm_b(
  Y,
  X_pos,
  Y_neg,
  Y_hk,
  Y_reg,
  cc,
  a,
  d_pos,
  d_neg,
  d_hk,
  d_reg,
  mu_b,
  phi,
  kappa_hk,
  kappa_reg,
  sig2_b,
  sig2_e,
  sig2_n,
  P,
  N,
  H,
  R,
  n_samples
)
}
\value{
A posterior sample of b
}
\description{
Update b
}
