% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_cc}
\alias{update_norm_cc}
\title{Update c}
\usage{
update_norm_cc(Y_neg, a, b, d_neg, sig2_n, L = -10, U = 10, N, n_samples)
}
\value{
A posterior sample of c
}
\description{
Update c
}
