% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_dhkreg}
\alias{update_norm_dhkreg}
\title{Update d^*_n}
\usage{
update_norm_dhkreg(Y, a, b, phi, kappa, sig2_e, sig2_d, n_genes, n_samples)
}
\value{
A posterior sample of d^*_n
}
\description{
Update d^*_n
}
