% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_dneg2}
\alias{update_norm_dneg2}
\title{Update d^-_n}
\usage{
update_norm_dneg2(Y_neg, cc, a, b, d_neg, sig2_dn, sig2_n, n_samples, N, j)
}
\value{
A posterior sample of d^-_n
}
\description{
Update d^-_n
}
