% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_dpos3}
\alias{update_norm_dpos3}
\title{Update d^+_n}
\usage{
update_norm_dpos3(
  Y,
  X_pos,
  a,
  b,
  sig2_d,
  sig2_e,
  P,
  n_samples,
  Sig_uns,
  D5,
  D6
)
}
\value{
A posterior sample of d^+_n
}
\description{
Update d^+_n
}
