% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_kappa}
\alias{update_norm_kappa}
\title{Update kappa}
\usage{
update_norm_kappa(
  Y_dat,
  a,
  b,
  phi,
  d,
  lambda,
  sig2_e,
  sig2_kappa,
  n_genes,
  n_samples
)
}
\value{
A posterior sample of kappa
}
\description{
Update kappa
}
