% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_mu}
\alias{update_norm_mu}
\title{Update mu}
\usage{
update_norm_mu(z, sig2_z, prior_mean, prior_var, L = -10, U = 10, n_samples)
}
\value{
A posterior sample of mu
}
\description{
Update mu
}
