% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_sig2_abd}
\alias{update_norm_sig2_abd}
\title{Update sigma^2_a, sigma^2_b, and sigma^2_d}
\usage{
update_norm_sig2_abd(z, mu_z, n, eps = 0.01)
}
\value{
A posterior sample of sigma^2_a, sigma^2_b, and sigma^2_d
}
\description{
Update sigma^2_a, sigma^2_b, and sigma^2_d
}
