% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_sig2_e}
\alias{update_norm_sig2_e}
\title{Update sigma^2_e}
\usage{
update_norm_sig2_e(
  Y_pos,
  X_pos,
  Y_hk,
  Y_reg,
  a,
  b,
  phi,
  kappa_hk,
  kappa_reg,
  d_pos,
  d_hk,
  d_reg,
  P,
  H,
  R,
  n_samples,
  eps = 0.01
)
}
\value{
A posterior sample of sigma^2_e
}
\description{
Update sigma^2_e
}
