% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_sig2_kappa}
\alias{update_norm_sig2_kappa}
\title{Update sigma^2_kappa}
\usage{
update_norm_sig2_kappa(kappa, lambda, n_genes, n_samples, eps = 0.01)
}
\value{
A posterior sample of sigma^2_kappa
}
\description{
Update sigma^2_kappa
}
