% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gibbs.R
\name{update_norm_sig2_n}
\alias{update_norm_sig2_n}
\title{Update sigma^2_e-}
\usage{
update_norm_sig2_n(Y_neg, a, b, cc, d_neg, n_genes, n_samples, eps = 0.01)
}
\value{
A posterior sample of sigma^2_e-
}
\description{
Update sigma^2_e-
}
