% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_gibbs.R
\name{update_s5k}
\alias{update_s5k}
\title{Update parameter s_5k}
\usage{
update_s5k(k, sjk)
}
\arguments{
\item{k}{The study index}

\item{sjk}{The current posterior sample of sjk}
}
\value{
A sample of the s5k parameter
}
\description{
This function performs one step of the Gibbs sampling
of the 5th components of the parameter s_jk
}
