% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_gibbs.R
\name{update_sig}
\alias{update_sig}
\title{Update parameter sigma_alpha/sigma_beta}
\usage{
update_sig(eps, alpha, mu, K = 13)
}
\arguments{
\item{eps}{The IG parameter}

\item{alpha}{The current posterior sample of alpha/beta}

\item{mu}{The current posterior sample of mu_alpha/mu_beta}

\item{K}{The number of studies}
}
\value{
A sample of the sigma_alpha/sigma_beta parameter
}
\description{
This function performs one step of the Gibbs sampling
of the parameter sigma_alpha/sigma_beta As the form of the updating rules
are identical for sigma_alpha and sigma_beta, we use the same
function for both of them
}
