% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_gibbs.R
\name{update_sigjk}
\alias{update_sigjk}
\title{Update parameter sigma^2_jk}
\usage{
update_sigjk(j, k, eps, ab, sjk)
}
\arguments{
\item{j}{The prob index}

\item{k}{The study index}

\item{eps}{The IG paramter}

\item{ab}{The current posterior sample of theta}

\item{sjk}{The current posterior sample of sjk}
}
\value{
A sample of the sigma^2_jk parameter
}
\description{
This function performs one step of the Gibbs sampling
of the parameter sigma^2_jk
}
