% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_gibbs.R
\name{update_sjk}
\alias{update_sjk}
\title{Update parameter s_jk}
\usage{
update_sjk(j, k, tj, tau2j, sigjk, ab, sjk)
}
\arguments{
\item{j}{The prob index}

\item{k}{The study index}

\item{tj}{The current posterior sample of t_j}

\item{tau2j}{The current posterior sample of tau^2_j}

\item{sigjk}{The current posterior sample of sigma^2_jk}

\item{ab}{The current posterior sample of theta}

\item{sjk}{The current posterior sample of sjk}
}
\value{
A sample of the sjk parameter
}
\description{
This function performs one step of the Gibbs sampling
of the first 4 components of the parameter s_jk
}
