% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_gibbs.R
\name{update_tau2j}
\alias{update_tau2j}
\title{Update parameter tau2j}
\usage{
update_tau2j(j, eps, sjk, tj, K = 13)
}
\arguments{
\item{j}{The prob index}

\item{eps}{The IG paramter}

\item{sjk}{The current posterior sample of sjk}

\item{tj}{The current posterior sample of t_j}

\item{K}{The number of studies}
}
\value{
A sample of the tau2j parameter
}
\description{
This function performs one step of the Gibbs sampling
of the parameter tau2j
}
