% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_gibbs.R
\name{update_tj}
\alias{update_tj}
\title{Update parameter tj}
\usage{
update_tj(j, L, U, sjk, tau2j, tj, K = 13)
}
\arguments{
\item{j}{The prob index}

\item{L}{The lower bound}

\item{U}{The upper bound}

\item{sjk}{The current posterior sample of sjk}

\item{tau2j}{The current posterior sample of tau^2_j}

\item{tj}{The current posterior sample of t_j}

\item{K}{The number of studies}
}
\value{
A sample of the tj parameter
}
\description{
This function performs one step of the Gibbs sampling
of the first 4 components of the parameter tj
}
