% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaNormData.R, R/normalization.R
\docType{data}
\name{MetaNorm}
\alias{MetaNorm}
\title{Datasets of MetaNorm}
\format{
An object of class \code{function} of length 1.
}
\usage{
MetaNorm

MetaNorm(
  dat,
  pos_conc = log10(c(128, 32, 8, 2, 0.5, 0.125)),
  M = 15000,
  n_keep = -1,
  mm = 3,
  seed = 0
)
}
\arguments{
\item{dat}{A list of 4 dataframes}

\item{pos_conc}{A vector of mRNA amount of positive probs}

\item{M}{Number of MCMC draws}

\item{n_keep}{Number of posterior draws to keep}

\item{mm}{The radius}

\item{seed}{Random seed}
}
\value{
A list containing posterior sample draws
}
\description{
Four .csv files are provided as example data on how to
utilize the MetaNorm package.

This function is similar to RCRnorm with the prior changed
to an informative prior infered from the meta analysis
}
\details{
meta_analysis_data.csv contains 13 datasets used for meta-analysis
which provided the MetaNorm function with an informative prior.

normalization_pos_data.csv, normalization_neg_data.csv,
normalization_hk_data.csv, normalization_reg_data.csv contain
mRNA measurements on 4 kinds of probes from one study. The data
was used by RCRnorm.

We seperated the original .RData file into
4 .csv files so that prosepective users can use them as a guidence
on how to organize their own data.

We also included 3 curated datasets based on meta_analysis_data.csv
with which users can compare to make sure that the output for their
data is of correct format.

The meta_analysis_data_coeff.csv contains our empirical estimates
of intercepts and slopes
}
\keyword{datasets}
