% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{MetaNorm}
\alias{MetaNorm}
\title{Estimate mRNA amount for every gene/sample combination}
\usage{
MetaNorm(
  dat,
  pos_conc = log10(c(128, 32, 8, 2, 0.5, 0.125)),
  M = 15000,
  n_keep = -1,
  mm = 3,
  seed = 0
)
}
\arguments{
\item{dat}{A list of 4 dataframes}

\item{pos_conc}{A vector of mRNA amount of positive probs}

\item{M}{Number of MCMC draws}

\item{n_keep}{Number of posterior draws to keep}

\item{mm}{The radius}

\item{seed}{Random seed}
}
\value{
A list containing posterior sample draws
}
\description{
This function is similar to RCRnorm with the prior changed
to an informative prior infered from the meta analysis
}
