% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaNormData.R, R/normalization.R
\docType{data}
\name{MetaNorm}
\alias{MetaNorm}
\title{Datasets of MetaNorm}
\format{
An object of class \code{function} of length 1.
}
\usage{
MetaNorm

MetaNorm(dat, M = 15000, n_keep = -1, mm = 3, seed = 0)
}
\arguments{
\item{dat}{A list of 4 dataframes}

\item{M}{Number of MCMC draws}

\item{n_keep}{Number of posterior draws to keep}

\item{mm}{The radius}

\item{seed}{Random seed}

\item{pos_conc}{A vector of mRNA amount of positive probs}
}
\value{
A list containing posterior sample draws
}
\description{
Several .Rdata files are provided as example data on how to
utilize the MetaNorm package.

This function is similar to RCRnorm with the prior changed
to an informative prior infered from the meta analysis
}
\details{
meta_analysis_data.RData contains 13 datasets used for meta-analysis
which provided the MetaNorm function with an informative prior.

normalization_data.RData contains
mRNA measurements on 4 kinds of probes from one study. The data
was used by RCRnorm.

meta_mcmc_draws1.RData, meta_mcmc_draw2.RData,
meta_mcmc_draws3.RData, meta_mcmc_draw4.RData
contain the posterior sample draws from 4
MCMC chains each with 12,000 samples from running meta_analysis.

norm_mcmc_draws_kappa_hk.RData contains the posterior sample draws from one
MCMC chains each with 2,000 samples from running MetaNorm
}
\keyword{datasets}
