% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metanorm.R
\name{metanormWorker}
\alias{metanormWorker}
\title{Perform normalization using a given compound's intensities.}
\usage{
metanormWorker(
  raw,
  order,
  keepScale,
  QConly,
  QCcheck,
  QCcheckp,
  changepoints,
  type,
  batch,
  batchwise,
  weights,
  model,
  k,
  cv,
  plotdir,
  plottype,
  i
)
}
\arguments{
\item{raw}{(numeric) Vector of length the number of samples, containing the
compounds' intensities.}

\item{order}{(numeric) Injection order of the samples. May be NULL if the
column order is chronological and no samples are omitted. If not
chronological, the argument's length should equal the number of samples.
Default NULL.}

\item{keepScale}{(logical) Whether the original intensity scale should be
retained, or whether rescaling can be performed. In the former case, fitted
values are first subtracted from observed values, and the median observed
value is re-added. In the latter case, normalization is achieved by dividing
by the fitted value, and values are centered around one. Default TRUE.}

\item{QConly}{(logical) Whether only QC samples should be used for fitting
the normalization model (FALSE), or whether also other samples contribute to
the model fit (TRUE). Sample contributions can be modified by the parameter
weights. Default FALSE.}

\item{QCcheck}{(logical) Whether to check for significantly different GAM
fits between QCs and samples (TRUE) or not (FALSE). In the case of
significantly different fits, QC samples are omitted and data normalized
using the sample data only. Default FALSE.}

\item{QCcheckp}{(numeric) Cutoff p-value for determining significantly 
different fits between QCs and samples separately, versus QCs and samples
combined. Default 0.1.}

\item{changepoints}{(logical) Whether to check for significant changepoints
within batches (TRUE) or not (FALSE). Only for tGAM/rGAM. Default FALSE.}

\item{type}{(character) Whether the sample is a QC (value 'QC') or not (any
other value). May be NULL, if QCOnly is set to FALSE. Default NULL.}

\item{batch}{(character) Vector of length the number of samples, indicating
for each sample to which batch it belongs. May be NULL, in which case samples
are presumed to belong to a single batch. Default NULL.}

\item{batchwise}{(logical) Whether to fit batchwise GAMs. Batchwise fitting
may result in significant speed gains for complex experiments with many
batches, especially when QCcheck is set to TRUE. Note that this changes the
QCcheck hypothesis test from a per-batch test (TRUE) to cross-batch tests
(FALSE). Default TRUE.}

\item{weights}{(numeric) Vector of length the number of samples, indicating
what weight should be attributed to each sample when fitting the
normalization model. Can be NULL, in which case all samples will be assigned
equal weight. Note: when fitting only to QC samples, the QCOnly approach
will be faster. Weights are only used for the rGAM and tGAM models, and
ignored otherwise. Default NULL.}

\item{model}{(character) One of 'rGAM', a robustified Gaussian GAM; 'tGAM',
a scaled-t GAM; 'QC-RLSC', as in Dunn et al. (2011); 'QC-RSC', as in Kirwan
et al. (2013); or 'rLOESS', a robustified version of QC-RLSC. Default: tGAM.}

\item{k}{(numeric) The maximum basis dimension to use when fitting GAMs.
Lower values result in faster analysis time. Calculated in the wrapper
function, as a function of two parameters.}

\item{cv}{(character) Type of cross-validation to use for QC-RLSC, QC-RSC and
rLOESS approaches. Either 'GCV' (faster) or 'LOOCV' (as suggested in the
original publications, but slower). Default: GCV.}

\item{plotdir}{(character) File location of plots (or NULL for no plots).}

\item{plottype}{(character) File type for writing plots, either "pdf" or 
"png", Default: pdf.}

\item{i}{(numeric) Compound (row) index, used for plot file naming.}
}
\value{
Normalized intensities for the compound.
}
\description{
Perform normalization using a given compound's intensities.
}
